Improve stubs
