+++ kryptone/kryptone-stubs/db/functions.pyi
@@ -1,0 +1,22 @@
+from kryptone.db.backends import SQLiteBackend
+
+
+class Functions:
+    backend: SQLiteBackend = ...
+
+    def __init__(self) -> None: ...
+
+    def functions_sql(self) -> str: ...
+
+
+class Lower(Functions):
+    field_name: str = ...
+
+    def __init__(self, field_name: str) -> None: ...
+    def __str__(self) -> str: ...
+
+    def function_sql(self) -> str: ...
+
+
+class Upper(Lower):
+    ...
